requirements.txt
asgiref==3.2.10
Django==3.0.8
django-admin-rangefilter==0.6.1
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
